# services-architecture/swagger/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Trabalho de conclusão de disciplina. \nDisciplina: Services Architecture / API / Mobile Architecture \nCurso: MBA em Engenharia de Software. \nTurma AOJ75\n\nResponsáveis: \n   Anelyn Ovalle | Fabio Alencar | Juliana Medeiros | Rafael Barbosa",
    "version": "1.0.0",
    "title": "Amazon API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost",
  "tags": [
    {
      "name": "product",
      "description": "cadastro, alteração e consulta dos produtos ofertados",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://swagger.io"
      }
    },
    {
      "name": "order",
      "description": "cadastro, consulta e alteração de pedidos",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://swagger.io"
      }
    },
    {
      "name": "ticket",
      "description": "cadastro e consulta de chamados",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://swagger.io"
      }
    }
  ],
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/product": {
      "post": {
        "tags": [
          "product"
        ],
        "summary": "adiciona um novo produto",
        "description": "",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "adicionar produto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/product/findById/{id}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "busca de produtos pelo Id",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "id": 10
              }
            },
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/product/wishlist": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "busca os produtos da lista de desejos",
        "description": "Produtos marcados serão exibidos nessa consulta.",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/product/wishlist/{id}": {
      "put": {
        "tags": [
          "product"
        ],
        "summary": "marca um produto como desejado",
        "description": "Produtos marcados serão exibidos na lista de desejos.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id do produto a ser marcado",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/components/requestBodies/ProductWishListBody"
            }
          }
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/product/findByGenre": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "busca de produtos pelo gênero",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "genre",
            "description": "busca pelo gênero do produto",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/product/findByMostViewed": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "busca de produtos pelo número de visualizações",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "number",
            "description": "número de visualizações",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/product/findByKeyWord": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "busca de produtos por palavra chave",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "description": "palavra chave",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/ticket": {
      "post": {
        "tags": [
          "ticket"
        ],
        "summary": "adiciona um novo chamado",
        "description": "",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Ticket"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/Ticket"
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/ticket/findById": {
      "get": {
        "tags": [
          "ticket"
        ],
        "summary": "busca de chamados pelo Id",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "ID do pedido",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Ticket"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/product/updatePrice/{id}": {
      "put": {
        "tags": [
          "product"
        ],
        "summary": "atualiza o preço de um produto",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id do produto a ter o preço atualizado.",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/components/requestBodies/ProductPriceUpdateBody"
            }
          }
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/ticket/findByType/{type}": {
      "get": {
        "tags": [
          "ticket"
        ],
        "summary": "busca de chamados pelo tipo",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "description": "busca pelo tipo do chamado",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Ticket"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/order": {
      "post": {
        "tags": [
          "order"
        ],
        "summary": "Create order",
        "operationId": "createOrder",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/components/requestBodies/CreatedOrder"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/order/{id}": {
      "get": {
        "tags": [
          "order"
        ],
        "summary": "consulta dos pedidos pelo id",
        "description": "",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "200",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/order/getListOrder": {
      "get": {
        "tags": [
          "order"
        ],
        "summary": "consulta a lista de todos os pedidos",
        "description": "",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Order"
              }
            }
          },
          "400": {
            "description": "Invalid tag value"
          }
        }
      }
    },
    "/order/calculateDeliveryPrice/{id}/{distance}": {
      "get": {
        "tags": [
          "order"
        ],
        "summary": "busca os produtos da lista de desejos",
        "description": "Cálculo do preço do frete pelo Id do pedido",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id do pedido submetido ao cálculo",
            "required": true,
            "type": "string"
          },
          {
            "name": "distance",
            "in": "path",
            "description": "distância em km",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "400": {
            "description": "Invalid tag value"
          }
        }
      }
    }
  },
  "definitions": {
    "OrderLine": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "productId": {
          "type": "integer",
          "format": "int64"
        },
        "quantity": {
          "type": "integer",
          "format": "int64"
        },
        "order_id": {
          "type": "integer",
          "format": "int64"
        }
      },
      "xml": {
        "name": "Order Line"
      }
    },
    "Product": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "readOnly": true
        },
        "name": {
          "type": "string",
          "example": "Monitor LG 24\""
        },
        "category": {
          "type": "string",
          "example": "Eletrônicos"
        },
        "price": {
          "type": "number",
          "example": 500
        },
        "genre": {
          "type": "string",
          "example": "Informática"
        },
        "description": {
          "type": "string",
          "example": "Monitor LED LG à venda."
        },
        "mostViewed": {
          "type": "integer",
          "format": "int64",
          "example": 0
        }
      },
      "xml": {
        "name": "Product"
      }
    },
    "Order": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "readOnly": true
        },
        "products": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "productid": {
                "type": "integer",
                "example": 1,
                "format": "intx64"
              },
              "quantity": {
                "type": "integer",
                "example": 1,
                "format": "intx64"
              }
            }
          }
        },
        "deliveryPrice": {
          "type": "number",
          "format": "0.00",
          "description": "Preço do frete em R$.",
          "readOnly": true
        },
        "status": {
          "type": "string",
          "readOnly": true,
          "example": "Criado"
        }
      },
      "xml": {
        "name": "Order"
      }
    },
    "Ticket": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "readOnly": true
        },
        "subject": {
          "type": "string",
          "example": "Pedido atrasado\""
        },
        "type": {
          "type": "string",
          "example": "Atraso de entrega"
        },
        "description": {
          "type": "string",
          "example": "Pedi um monitor com previsão de entrega para a semana passada que ainda não chegou."
        }
      },
      "xml": {
        "name": "Product"
      }
    }
  },
  "components": {
    "requestBodies": {
      "ProductPriceUpdateBody": {
        "required": true,
        "content": {
          "application/json": null
        },
        "properties": {
          "price": {
            "type": "number",
            "example": 600
          }
        }
      },
      "ProductWishListBody": {
        "required": true,
        "content": {
          "application/json": null
        },
        "properties": {
          "wishList": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "CreatedOrder": {
        "required": true,
        "content": {
          "application/json": null
        },
        "properties": {
          "retorno": {
            "type": "string",
            "example": "Status do pedido = {$status}. O total do pedido é: {$totalPedido}"
          }
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}